This updates this branch with trunk.  Next up, initial work to merge the OpenMP scheduler with logic to prepare simulation variables before task execution
Former-commit-id: 6b2972ef6f5a668f6ebb779a5ba4ebe5ed9bfe2e

+++ src/StandAlone/inputs/ARCHES/kokkos_solver_tests/Verification/GridIndStudyTemp.py
@@ -139,6 +139,9 @@ def main():
   parser.add_argument('-var_mms', required=True,
                       help='name of mms')
 
+  parser.add_argument('-keep_uda', 
+                      help='Keep the udas - do not delete them.', action='store_true')
+
   args = parser.parse_args()
   
   # if the number of levels is not provided, set it to 3
@@ -404,7 +407,8 @@ def main():
        plt.savefig(data+'/'+basename)
 
   plt.show()
-  os.system('rm -rf *.uda*')
+  if args.keep_uda is None:
+    os.system('rm -rf *.uda*')
   os.system('rm -rf *.dot')
   os.system('rm log.txt')  
   #os.system('rm data/*.txt')
